Refactor: move profit code to profit module; split crafting module to focus on crafting
