+++ src/crafting.rs
@@ -3,6 +3,7 @@ use crate::config;
 use crate::recipe::Recipe;
 use crate::item::Item;
 use crate::money::Money;
+use crate::profit;
 
 use num_rational::Ratio;
 use num_traits::Zero;
@@ -89,14 +90,14 @@ pub fn calculate_estimated_min_crafting_cost(
 // Exact
 
 #[derive(Debug, Copy, Clone)]
-struct PreciseCraftingCost {
-    cost: Money,
-    source: Source,
+pub struct PreciseCraftingCost {
+    pub cost: Money,
+    pub source: Source,
 }
 
-struct PreciseCraftingCostContext {
-    purchases: Vec<(u32, u32, Source)>, // id, count, Source
-    items: CraftedItems,
+pub struct PreciseCraftingCostContext {
+    pub purchases: Vec<(u32, u32, Source)>, // id, count, Source
+    pub items: CraftedItems,
 }
 
 #[derive(Debug, Default, Clone, Eq, PartialEq)]
@@ -106,7 +107,7 @@ pub struct CraftedItems {
 }
 
 impl CraftedItems {
-    fn crafting_steps(
+    pub fn crafting_steps(
         &self,
         recipes_map: &HashMap<u32, Recipe>,
     ) -> Ratio<u32> {
@@ -206,12 +207,12 @@ impl CraftedItems {
 
 // Calculate the lowest cost method to obtain the given item, with simulated purchases from
 // the trading post.
-fn calculate_precise_min_crafting_cost(
+pub fn calculate_precise_min_crafting_cost(
     item_id: u32,
     item_count: u32,
     recipes_map: &HashMap<u32, Recipe>,
     items_map: &HashMap<u32, Item>,
-    tp_listings_map: &mut BTreeMap<u32, ItemListings>,
+    tp_listings_map: &mut BTreeMap<u32, profit::ItemListings>,
     context: &mut PreciseCraftingCostContext,
     opt: &config::CraftingOptions,
 ) -> Option<PreciseCraftingCost> {
@@ -364,158 +365,6 @@ fn calculate_precise_min_crafting_cost(
     Some(PreciseCraftingCost { cost: cost + cost_of_leftovers_used, source })
 }
 
-pub fn calculate_crafting_profit(
-    item_id: u32,
-    recipes_map: &HashMap<u32, Recipe>,
-    items_map: &HashMap<u32, Item>,
-    tp_listings_map: &HashMap<u32, api::ItemListings>,
-    mut purchased_ingredients: Option<&mut HashMap<(u32, Source), PurchasedIngredient>>,
-    opt: &config::CraftingOptions,
-) -> Option<ProfitableItem> {
-    let mut tp_listings_map: BTreeMap<u32, ItemListings> = tp_listings_map
-        .clone()
-        .into_iter()
-        .map(|(id, listings)| (id, ItemListings::from(listings)))
-        .collect();
-
-    let recipe = recipes_map.get(&item_id);
-    let output_item_count = recipe.map(|recipe| recipe.output_item_count).unwrap_or(1);
-    let threshold = Money::from_copper(opt.threshold.unwrap_or(0) as i32);
-
-    let mut listing_profit = Money::zero();
-    let mut total_crafting_cost = Money::zero();
-    let mut crafting_count = 0;
-    let mut crafted_items = CraftedItems::default();
-
-    let mut min_sell = 0;
-    let max_sell = tp_listings_map
-        .get(&item_id)
-        .map_or_else(|| opt.threshold.unwrap_or(0), |listings| {
-            listings
-                .buys
-                .last()
-                .map_or(0, |l| l.unit_price)
-        });
-    let mut breakeven = Money::zero();
-
-    // simulate crafting 1 item per loop iteration until it becomes unprofitable
-    loop {
-        if let Some(count) = opt.count {
-            if crafting_count + output_item_count > count {
-                break;
-            }
-        }
-
-        let mut context = PreciseCraftingCostContext {
-            purchases: vec![],
-            items: crafted_items.clone(),
-        };
-
-        let crafting_cost = if let Some(PreciseCraftingCost {
-            source: Source::Crafting,
-            cost,
-        }) = calculate_precise_min_crafting_cost(
-            item_id,
-            output_item_count,
-            recipes_map,
-            items_map,
-            &mut tp_listings_map,
-            &mut context,
-            opt,
-        ) {
-            cost
-        } else {
-            break;
-        };
-
-
-        let (buy_price, min_buy) = if let Some(price) = opt.value {
-            (Money::from_copper(price as i32), price)
-        } else if let Some((buy_price, min_buy)) = tp_listings_map
-            .get_mut(&item_id)
-            .unwrap_or_else(|| panic!("Missing listings for item id: {}", item_id))
-            .sell(output_item_count)
-        {
-            (buy_price, min_buy)
-        } else {
-            break;
-        };
-
-        // Ensure buy_price is larger before subtracting cost for profit
-        if buy_price < crafting_cost + threshold {
-            break;
-        }
-
-        listing_profit += buy_price - crafting_cost;
-        total_crafting_cost += crafting_cost;
-        crafting_count += output_item_count;
-        crafted_items = context.items;
-
-        min_sell = min_buy;
-        // Breakeven is based on the last/most expensive to craft
-        breakeven = crafting_cost / output_item_count;
-
-        // Finalize purchases
-        for (purchase_id, count, purchase_source) in &context.purchases {
-            let (cost, min_sell, max_sell) = if let Source::TradingPost = *purchase_source {
-                let listing = tp_listings_map.get_mut(purchase_id).unwrap_or_else(|| {
-                    panic!(
-                        "Missing listings for ingredient {} of item id {}",
-                        purchase_id, item_id
-                    )
-                });
-                listing.pending_buy_quantity -= *count;
-                let (cost, min_sell, max_sell) = listing.buy(*count).unwrap_or_else(|| {
-                    panic!(
-                        "Expected to be able to buy {} of ingredient {} for item id {}",
-                        count, purchase_id, item_id
-                    )
-                });
-                (cost, min_sell, max_sell)
-            } else {
-                (0, 0, 0)
-            };
-
-            if let Some(purchased_ingredients) = &mut purchased_ingredients {
-                let ingredient = purchased_ingredients
-                    .entry((*purchase_id, *purchase_source))
-                    .or_insert_with(|| PurchasedIngredient {
-                        count: 0,
-                        max_price: Money::default(),
-                        min_price: Money::default(),
-                        total_cost: Money::default(),
-                    });
-                ingredient.count += count;
-                if ingredient.min_price.is_zero() {
-                    ingredient.min_price = Money::from_copper(min_sell as i32);
-                }
-                ingredient.max_price = Money::from_copper(max_sell as i32);
-                ingredient.total_cost += Money::from_copper(cost as i32);
-            }
-        }
-        debug_assert!(tp_listings_map
-            .iter()
-            .all(|(_, listing)| listing.pending_buy_quantity == 0));
-
-    }
-
-    if crafting_count > 0 && !listing_profit.is_zero() {
-        Some(ProfitableItem {
-            id: item_id,
-            crafting_cost: total_crafting_cost,
-            profit: listing_profit,
-            count: crafting_count,
-            max_sell: Money::from_copper(max_sell as i32),
-            min_sell: Money::from_copper(min_sell as i32),
-            breakeven: breakeven.trading_post_listing_price(),
-            crafting_steps: crafted_items.crafting_steps(recipes_map).to_integer(),
-            crafted_items,
-        })
-    } else {
-        None
-    }
-}
-
 #[derive(Debug, Eq, PartialEq)]
 pub struct PurchasedIngredient {
     pub count: u32,
@@ -524,165 +373,6 @@ pub struct PurchasedIngredient {
     pub total_cost: Money,
 }
 
-#[derive(Debug, Eq, PartialEq)]
-pub struct ProfitableItem {
-    pub id: u32,
-    pub crafting_cost: Money,
-    pub count: u32,
-    pub profit: Money,
-    pub max_sell: Money,
-    pub min_sell: Money,
-    pub breakeven: Money,
-    pub crafting_steps: u32,
-    pub crafted_items: CraftedItems,
-}
-
-impl ProfitableItem {
-    pub fn profit_per_item(&self) -> Money {
-        self.profit / self.count
-    }
-
-    pub fn profit_per_crafting_step(&self) -> Money {
-        self.profit / self.crafting_steps
-    }
-
-    pub fn profit_on_cost(&self) -> f64 {
-        self.profit.percent(self.crafting_cost)
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct ItemListings {
-    pub id: u32,
-    pub buys: Vec<Listing>,
-    pub sells: Vec<Listing>,
-    pub pending_buy_quantity: u32,
-}
-
-#[derive(Clone, Debug)]
-pub struct Listing {
-    pub unit_price: u32,
-    pub quantity: u32,
-}
-
-impl ItemListings {
-    fn buy(&mut self, mut count: u32) -> Option<(u32, u32, u32)> {
-        let mut cost = 0;
-        let mut min_sell = 0;
-        let mut max_sell = 0;
-
-        while count > 0 {
-            // sells are sorted in descending price
-            let remove = if let Some(listing) = self.sells.last_mut() {
-                listing.quantity -= 1;
-                count -= 1;
-                if min_sell == 0 {
-                    min_sell = listing.unit_price;
-                }
-                max_sell = listing.unit_price;
-                cost += listing.unit_price;
-                listing.quantity.is_zero()
-            } else {
-                return None;
-            };
-
-            if remove {
-                self.sells.pop();
-            }
-        }
-
-        Some((cost, min_sell, max_sell))
-    }
-
-    fn sell(&mut self, mut count: u32) -> Option<(Money, u32)> {
-        let mut revenue = Money::zero();
-        let mut min_buy = 0;
-
-        while count > 0 {
-            // buys are sorted in ascending price
-            let remove = if let Some(listing) = self.buys.last_mut() {
-                listing.quantity -= 1;
-                count -= 1;
-                min_buy = listing.unit_price;
-                revenue += Money::from_copper(listing.unit_price as i32).trading_post_sale_revenue();
-                listing.quantity.is_zero()
-            } else {
-                return None;
-            };
-
-            if remove {
-                self.buys.pop();
-            }
-        }
-
-        Some((revenue, min_buy))
-    }
-
-    fn lowest_sell_offer(&self, mut quantity: u32) -> Option<u32> {
-        debug_assert!(!quantity.is_zero());
-
-        let mut cost = 0;
-        let mut pending_buy_quantity = self.pending_buy_quantity;
-
-        for listing in self.sells.iter().rev() {
-            let mut remaining_listing_quantity = listing.quantity;
-            if pending_buy_quantity > 0 {
-                if pending_buy_quantity >= remaining_listing_quantity {
-                    pending_buy_quantity -= remaining_listing_quantity;
-                    remaining_listing_quantity = 0;
-                } else {
-                    remaining_listing_quantity -= pending_buy_quantity;
-                    pending_buy_quantity = 0;
-                }
-            }
-
-            if remaining_listing_quantity > 0 {
-                if remaining_listing_quantity < quantity {
-                    quantity -= remaining_listing_quantity;
-                    cost += remaining_listing_quantity * listing.unit_price;
-                } else {
-                    cost += quantity * listing.unit_price;
-                    quantity = 0;
-                }
-            }
-
-            if quantity.is_zero() {
-                break;
-            }
-        }
-
-        if quantity > 0 {
-            None
-        } else {
-            Some(cost)
-        }
-    }
-}
-
-impl From<api::ItemListings> for ItemListings {
-    fn from(v: api::ItemListings) -> Self {
-        ItemListings {
-            id: v.id,
-            buys: v
-                .buys
-                .into_iter()
-                .map(|listing| Listing {
-                    unit_price: listing.unit_price,
-                    quantity: listing.quantity.into(),
-                })
-                .collect(),
-            sells: v
-                .sells
-                .into_iter()
-                .map(|listing| Listing {
-                    unit_price: listing.unit_price,
-                    quantity: listing.quantity.into(),
-                })
-                .collect(),
-            pending_buy_quantity: 0,
-        }
-    }
-}
 
 trait OptionInnerMin<T> {
     fn inner_min(self, other: Option<T>) -> Option<T>;
